Give Subscription its own source file
